Fixed bug #408 - service checks get duplicated on reload.
If a check is still running when Nagios is restarted or reloaded (via SIGHUP),
Nagios will now look for running checks and mark them as executing. Then, when
the checks are scheduled before the execution loop starts, the running checks
will be not be rescheduled. They will be rescheduled as usual when their 
results are reaped.

This commit also re-introduces the extra event debug logging that was 
unintentionally removed between versions 3.4.2 and 3.4.3.


git-svn-id: https://nagios.svn.sourceforge.net/svnroot/nagios/nagioscore/branches/nagios-3-4-x@2554 5f96b256-904b-4d8d-8c98-d829582c6739

+++ base/utils.c
@@ -2284,7 +2284,8 @@ int process_check_result_queue(char *dirname) {
 				continue;
 
 			/* process the file */
-			result = process_check_result_file(file, &check_result_list);
+			result = process_check_result_file(file, &check_result_list, 
+					TRUE, TRUE);
 
 			/* break out if we encountered an error */
 			if(result == ERROR)
@@ -2301,14 +2302,142 @@ int process_check_result_queue(char *dirname) {
 
 
 
+/* Find checks that are currently executing. This function is intended to 
+   be used on a Nagios restart to prevent currently executing checks from
+   being rescheduled. */
+int find_executing_checks(char *dirname) {
+	char file[MAX_FILENAME_LENGTH];
+	DIR *dirp = NULL;
+	struct dirent *dirfile = NULL;
+	int x = 0;
+	struct stat stat_buf;
+	int result = OK;
+	check_result *crl = NULL;
+	check_result *cr = NULL;
+	service *svc = NULL;
+	host *host = NULL;
+	time_t current_time;
+
+	log_debug_info(DEBUGL_FUNCTIONS, 0, "find_executing_checks() start\n");
+
+	/* make sure we have what we need */
+	if(dirname == NULL) {
+		logit(NSLOG_CONFIG_ERROR, TRUE, "Error: No check directory specified.\n");
+		return ERROR;
+		}
+
+	/* open the directory for reading */
+	if((dirp = opendir(dirname)) == NULL) {
+		logit(NSLOG_CONFIG_ERROR, TRUE, "Error: Could not open check directory '%s' for reading.\n", dirname);
+		return ERROR;
+		}
+
+	log_debug_info(DEBUGL_CHECKS, 1, "Starting to read check directory '%s'...\n", dirname);
+
+	/* process all files in the directory... */
+	while((dirfile = readdir(dirp)) != NULL) {
+
+		/* create /path/to/file */
+		snprintf(file, sizeof(file), "%s/%s", dirname, dirfile->d_name);
+		file[sizeof(file) - 1] = '\x0';
+
+		/* process this if it's a check result file... */
+		x = strlen(dirfile->d_name);
+		if(x == 11 && !strncmp( dirfile->d_name, "check", 5)) {
+
+			if(stat(file, &stat_buf) == -1) {
+				logit(NSLOG_RUNTIME_WARNING, TRUE, "Warning: Could not stat() check status '%s'.\n", file);
+				continue;
+				}
+
+			switch(stat_buf.st_mode & S_IFMT) {
+
+				case S_IFREG:
+					/* don't process symlinked files */
+					if(!S_ISREG(stat_buf.st_mode))
+						continue;
+					break;
+
+				default:
+					/* everything else we ignore */
+					continue;
+					break;
+				}
+
+			/* at this point we have a regular file... */
+			log_debug_info(DEBUGL_CHECKS, 2, 
+					"Looking for still-executing checks in %s.\n", file);
+
+			/* process the file */
+			result = process_check_result_file(file, &crl, FALSE, FALSE);
+
+			/* break out if we encountered an error */
+			if(result == ERROR)
+				break;
+
+			time(&current_time);
+
+			/* examine the check results */
+			while(( cr = read_check_result(&crl)) != NULL) {
+				if( HOST_CHECK == cr->object_check_type) {
+					log_debug_info(DEBUGL_CHECKS, 2, 
+							"Determining whether check for host '%s' is still executing.\n",
+							cr->host_name);
+					if((host = find_host(cr->host_name)) == NULL) {
+						logit(NSLOG_RUNTIME_WARNING, TRUE, 
+								"Warning: Check status contained host '%s', "
+								"but the host could not be found! Ignoring "
+								"check.\n", cr->host_name);
+						}
+					if( current_time - cr->start_time.tv_sec < 
+							host_check_timeout) {
+						log_debug_info(DEBUGL_CHECKS, 1, 
+								"Check for host %s is still executing.\n",
+								cr->host_name);
+						host->is_executing = TRUE;
+						}
+					}
+				else if( SERVICE_CHECK == cr->object_check_type) {
+					log_debug_info(DEBUGL_CHECKS, 2, 
+							"Determining whether check for service '%s' on host '%s' isstill executing.\n",
+							cr->host_name, cr->service_description);
+					if((svc = find_service(cr->host_name, 
+							cr->service_description)) == NULL) {
+						logit(NSLOG_RUNTIME_WARNING, TRUE, 
+								"Warning: Check status contained service '%s' "
+								"on host '%s', but the service could not be "
+								"found! Ignoring check.\n", 
+								cr->service_description, cr->host_name);
+						}
+					if( current_time - cr->start_time.tv_sec < 
+							service_check_timeout) {
+						log_debug_info(DEBUGL_CHECKS, 1, 
+								"Check for service %s:%s is still executing.\n",
+								cr->host_name, cr->service_description);
+						svc->is_executing = TRUE;
+						}
+					}
+				free_check_result_list(&crl);
+				}
+			}
+		}
+
+	closedir(dirp);
+
+	return result;
+
+	}
+
+
+
+
 /* reads check result(s) from a file */
-int process_check_result_file(char *fname, check_result **listp) {
+int process_check_result_file(char *fname, check_result **listp, int delete_file, int need_output) {
 	mmapfile *thefile = NULL;
 	char *input = NULL;
 	char *var = NULL;
 	char *val = NULL;
 	char *v1 = NULL, *v2 = NULL;
-	int delete_file = FALSE;
 	time_t current_time;
 	check_result *new_cr = NULL;
 
@@ -2349,7 +2478,8 @@ int process_check_result_file(char *fname, check_result **listp) {
 			if(new_cr) {
 
 				/* do we have the minimum amount of data? */
-				if(new_cr->host_name != NULL && new_cr->output != NULL) {
+				if(new_cr->host_name != NULL && 
+						( !need_output || new_cr->output != NULL)) {
 
 					/* add check result to list in memory */
 					add_check_result_to_list(listp, new_cr);
@@ -2444,7 +2574,8 @@ int process_check_result_file(char *fname, check_result **listp) {
 	if(new_cr) {
 
 		/* do we have the minimum amount of data? */
-		if(new_cr->host_name != NULL && new_cr->output != NULL) {
+		if(new_cr->host_name != NULL && 
+				( !need_output || new_cr->output != NULL)) {
 
 			/* add check result to list in memory */
 			add_check_result_to_list(listp, new_cr);
@@ -2465,9 +2596,9 @@ int process_check_result_file(char *fname, check_result **listp) {
 	my_free(input);
 	mmap_fclose(thefile);
 
-	/* delete the file (as well its ok-to-go file) if it's too old */
-	/* other (current) files are deleted later (when results are processed) */
-	delete_check_result_file(fname);
+	/* delete the file (as well its ok-to-go file) if they exist */
+	if(TRUE == delete_file)
+		delete_check_result_file(fname);
 
 	return OK;
 	}
